+++ features/steps/notifications_steps.rb
@@ -62,6 +62,18 @@ Given /^the user wants to receive Voiceblue notifications for entity '([\w\.\-]+
                               :redis => @notifier_redis )
 end
 
+Given /^the user wants to receive Hipchat notifications for entity '([\w\.\-]+)'$/ do |entity|
+  add_contact( 'id'         => '0999',
+               'first_name' => 'John',
+               'last_name'  => 'Smith',
+               'email'      => 'johns@example.dom',
+               'media'      => {'hipchat' => 'room1'} )
+  Flapjack::Data::Entity.add({'id'       => '5000',
+                              'name'     => entity,
+                              'contacts' => ["0999"]},
+                              :redis => @notifier_redis )
+end
+
 Given /^the user wants to receive email notifications for entity '([\w\.\-]+)'$/ do |entity|
   add_contact( 'id'         => '0999',
                'first_name' => 'John',
@@ -135,6 +147,11 @@ end
 
 Then /^an Voiceblue notification for entity '([\w\.\-]+)' should be queued for the user$/ do |entity|
   queue = redis_peek('voiceblue_email2sms_notifications')
+  expect(queue.select {|n| n['event_id'] =~ /#{entity}:ping/ }).not_to be_empty
+end
+
+Then /^a Hipchat notification for entity '([\w\.\-]+)' should be queued for the user$/ do |entity|
+  queue = redis_peek('hipchat_notifications')
   expect(queue.select {|n| n['event_id'] =~ /#{entity}:ping/ }).not_to be_empty
 end
 
@@ -238,6 +255,26 @@ Given /^a user Voiceblue notification has been queued for entity '([\w\.\-]+)'$/
                             :redis => @notifier_redis)
 end
 
+Given /^a user Hipchat notification has been queued for entity '([\w\.\-]+)'$/ do |entity|
+  Flapjack::Data::Entity.add({'id'   => '5000',
+                              'name' => entity},
+                             :redis => @redis )
+  @hipchat_notification = {'notification_type'  => 'problem',
+                       'contact_first_name' => 'John',
+                       'contact_last_name'  => 'Smith',
+                       'state'              => 'critical',
+                       'summary'            => 'Socket timeout after 10 seconds',
+                       'time'               => Time.now.to_i,
+                       'event_id'           => "#{entity}:ping",
+                       'address'            => 'room1',
+                       'id'                 => 1,
+                       'state_duration'     => 30,
+                       'duration'           => 45}
+
+  Flapjack::Data::Alert.add('hipchat_notifications', @hipchat_notification,
+                            :redis => @notifier_redis)
+end
+
 Given /^a user email notification has been queued for entity '([\w\.\-]+)'$/ do |entity|
   Flapjack::Data::Entity.add({'id'   => '5001',
                               'name' => entity},
@@ -311,6 +348,18 @@ When /^the Voiceblue notification handler runs successfully$/ do
   drain_alerts('voiceblue_email2sms_notifications', @voiceblue_email)
 end
 
+When /^the Hipchat notification handler runs successfully$/ do
+  # Stub Request
+  stub_request(:post, "https://api.hipchat.com/v2/room/room1/notification?auth_token=TOKEN123").
+        to_return(:status => 200)
+  
+  @hipchat_notification = Flapjack::Gateways::Hipchat.new(:config => {
+    'auth_token' => 'TOKEN123',  
+  }, :redis_config => @redis_opts, :logger => @logger)
+
+  drain_alerts('hipchat_notifications', @hipchat_notification)
+end
+
 When /^the SMS notification handler fails to send an SMS$/ do
   @request = stub_request(:get, /^#{Regexp.escape(Flapjack::Gateways::SmsMessagenet::MESSAGENET_DEFAULT_URL)}/).to_return(:status => [500, "Internal Server Error"])
 
@@ -349,6 +398,18 @@ When /^the Voiceblue notification handler fails to send an SMS$/ do
   }, :redis_config => @redis_opts, :logger => @logger)
 
   drain_alerts('voiceblue_email2sms_notifications', @voiceblue_email)
+end
+
+When /^the Hipchat notification handler fails to send a notification$/ do
+  # Stub Request
+  stub_request(:post, "https://api.hipchat.com/v2/room/room1/notification?auth_token=TOKEN123").
+        to_return(:status => 500)
+
+  @hipchat_notification = Flapjack::Gateways::Hipchat.new(:config => {
+    'auth_token' => 'TOKEN123',
+  }, :redis_config => @redis_opts, :logger => @logger)
+
+  drain_alerts('hipchat_notifications', @hipchat_notification)
 end
 
 When /^the email notification handler runs successfully$/ do
@@ -405,6 +466,10 @@ end
 
 Then /^the user should( not)? receive an Voiceblue notification$/ do |negativity|
   expect(@voiceblue_email.instance_variable_get('@sent')).to eq(negativity.nil? ? 1 : 0)
+end
+
+Then /^the user should( not)? receive a Hipchat notification$/ do |negativity|
+  expect(@hipchat_notification.instance_variable_get('@sent')).to eq(negativity.nil? ? 1 : 0)
 end
 
 Then /^the user should( not)? receive an email notification$/ do |negativity|