+++ features/support/env.rb
@@ -60,6 +60,10 @@ class MockEmailer
   include EM::Deferrable
 end
 
+class MockWeb
+  include EM::Deferrable
+end
+
 class RedisDelorean
 
   ShiftTTLsScript = <<SHIFT_TTLS
@@ -172,6 +176,7 @@ Before('@notifier') do
                 'sms_nexmo_queue' => 'sms_nexmo_notifications',
                 'sns_queue' => 'sns_notifications',
                 'voiceblue_email2sms_queue' => 'voiceblue_email2sms_notifications',
+                'hipchat_queue' => 'hipchat_notifications',
                 'default_contact_timezone' => 'America/New_York'})
   @notifier_redis = @notifier.instance_variable_get('@redis')
 end