+++ lib/flapjack/notifier.rb
@@ -20,6 +20,7 @@ require 'flapjack/gateways/sms_twilio'
 require 'flapjack/gateways/sms_nexmo'
 require 'flapjack/gateways/aws_sns'
 require 'flapjack/gateways/voiceblue_email2sms'
+require 'flapjack/gateways/hipchat'
 
 module Flapjack